+++ index.js
@@ -3,6 +3,7 @@
 module.exports = {
   rules: {
     'no-let': require('./rules/no-let'),
+    'no-mutating-assign': require('./rules/no-mutating-assign'),
     'no-mutation': require('./rules/no-mutation'),
     'no-this': require('./rules/no-this')
   },
@@ -11,6 +12,7 @@ module.exports = {
       rules: {
         'no-var': 'error',
         'fp/no-let': 'error',
+        'fp/no-mutating-assign': 'error',
         'fp/no-mutation': 'error',
         'fp/no-this': 'error'
       }